add unit tests for ResourceExtension -> generateMissingItem class
